requirements.txt
Flask_Script==2.0.6
Flask_MySQLdb==0.2.0
Flask_SQLAlchemy==2.4.0
SQLAlchemy==1.3.4
Flask_Migrate==2.5.2
alembic==1.0.10
Flask==1.0.3
python-dotenv==0.10.3
psycopg2==2.8.2
gunicorn==19.0.0
